refactor: extract core PMS migrations
